validation in category action v2.0.0
