+++ FrontEnd/src/pages/AdminPage/FormatCategories/CategoriesTable.jsx
