+++ FrontEnd/src/pages/AdminPage/FormatCategories/CategoriesTable.module.scss
