+++ FrontEnd/src/pages/AdminPage/FormatCategories/CategoryAdd.jsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import { Button, Input } from 'antd';
 import { toast } from 'react-toastify';
+import ValidateCategory from './CategoryValidation';
 import 'react-toastify/dist/ReactToastify.css';
 import axios from 'axios';
 import PropTypes from 'prop-types';
@@ -10,49 +11,11 @@ function CategoryAdd({ onActionComplete }) {
     const [categoryAddName, setCategoryAddName] = useState('');
     const [isAdded, setIsAdded] = useState(false);
     const [error, setError] = useState('');
-
-    const validateCategory = async (category) => {
-        const trimmedCategory = category.trim();
-        const regex = /^[А-ЯЇЄҐ][а-яїієґА-ЯЇІЄҐ\s]*$/;
-
-        if (trimmedCategory.length < 2 || trimmedCategory.length > 50) {
-            setError('Назва категорії має бути від 2 до 50 символів.');
-            return false;
-        }
-
-        if (!regex.test(trimmedCategory)) {
-            setError('Назва категорії має починатися з великої літери та містити лише кириличні символи.');
-            return false;
-        }
-
-        try {
-            const response = await axios.get(
-                `${process.env.REACT_APP_BASE_API_URL}/api/admin/categories/?name=${encodeURIComponent(trimmedCategory)}`
-            );
-
-            const categoryExists = response.data.results.some(
-                (item) => item.name.toLowerCase() === trimmedCategory.toLowerCase()
-            );
-
-            if (categoryExists) {
-                setError('Така категорія вже існує.');
-                return false;
-            }
-        } catch (error) {
-            setError('Помилка перевірки. Спробуйте пізніше.');
-            return false;
-        }
-
-        setError('');
-        return true;
-    };
-
-
     const handleCategoryAdd = async () => {
         if (isAdded) return;
 
         setIsAdded(true);
-        const isValid = await validateCategory(categoryAddName);
+        const isValid = await ValidateCategory(categoryAddName, setError);
         if (!isValid) {
             setIsAdded(false);
             return;
@@ -67,7 +30,6 @@ function CategoryAdd({ onActionComplete }) {
             setCategoryAddName('');
             if (onActionComplete) onActionComplete();
         } catch (error) {
-            console.error('Error adding category:', error);
             toast.error('Не вдалося створити категорію.');
         } finally {
             setIsAdded(false);
@@ -85,7 +47,7 @@ function CategoryAdd({ onActionComplete }) {
                 onChange={(e) => {
                     const input = e.target.value;
                     setCategoryAddName(input);
-                    validateCategory(input);
+                    ValidateCategory(input, setError);
                 }}
                 className={styles.CategoryAddTextarea}
             />
